[roll] Roll fuchsia [gidl][benchmarks] Byte array and struct field benchmarks
Bindings that perform copies to encode structs should have similar
performance for byte array and struct fields.

Byte array sizes adjusted to align with struct field benchmarks to be
easier to compare (at the low end) and to test performance of encoding a
max-sized array for a channel (at the high end).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380014
GitOrigin-RevId: 71f9464241764548e973db7bea79d158ad2da2a8
Change-Id: I08ee4a7e34cb4f52fdc189a1b4e161102c04a28b