[roll] Roll fuchsia [microbenchmarks] Do multiple process runs in SL4F test for parity with non-SL4F
This is in preparation for switching the perfcompare trybot to using
the SL4F-based wrapper
(src/tests/end_to_end/perf/test/microbenchmarks_test.dart) instead of
the older non-SL4F-based wrapper (src/tests/benchmarks/benchmarks.cc).

Change microbenchmarks_test.dart to do multiple process runs of
fuchsia_microbenchmarks similar to how benchmarks.cc does that.  The
reasons are described in the comments, which are based on the comments
in benchmarks.cc.

A difference from benchmarks.cc is that we do the extra process runs
unconditionally, even when the test is not running in the context of
perfcompare.  This is for simplicity, and the extra running time is
not much.  (There is no flag plumbed through to indicate whether the
test is running on a perfcompare builder.  We could check the
BUILDER_NAME env var, but we don't really want to conditionalise on
infra builder names.)

Original-Bug: 49280
Test: on CQ: fuchsia-x64-perfcompare-dawson_canyon-sl4f
 * Confidence intervals *should* get narrower, but this is currently
   confounded somewhat by issue 49784.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380082
GitOrigin-RevId: df0895e312060157ce04a9dc6f392ea4d792f611
Change-Id: Ib033825a924cdbb6afe92205df901faa7405c2ad