[roll] Roll fuchsia [sshutil] Have ssh commands be slices of strings
sshutil previously represented commands to run over ssh as strings with
whitespace-separated arguments. This is the form in which the Go ssh
library takes commands, but it's inconsistent with Go stdlib functions
like exec.Command(), as well as the SSHRunner library that botanist and
testrunner use.

I also reformatted many Run() calls of the form:

  err := client.Run("command")
  if err != nil { ... }

to be of the form:

  cmd := []string{"command"}
  if err := client.Run(cmd); err != nil { ... }

This is consistent with the Go style guide. I kept `cmd` declarations on
a separate line to keep the `if` line relatively short.

Original-Bug: 48042
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380073
GitOrigin-RevId: 1d35171be7abfe8e785f0840f6805a9884ea2200
Change-Id: If2363b2dedab9da4c8b8aec89bd15b95da54c160