[roll] Roll fuchsia [component_manager] Refactor RightWalkState into a generic struct to reuse for filters
This struct is quite generic except for the comparison on the rights.
Refactored rights specific things into an impl of WalkStateUnit.

This should enabling reusing this logic for Filter routing which looks
very similar to rights routing, but with a different way of checking
if a next state is valid.

Original-Bug: 49553
Tested: fx test component_manager_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379317
GitOrigin-RevId: b6a30572c61f7b341256c6de07106300d2306f3f
Change-Id: I11266f037f5b92c5e82c8f96407a0cb10b0cb8eb