[roll] Roll fuchsia [scenic] Enable Vulkan pipeline cache warming in Scenic.
Step 9 in a series of CLs to build Vulkan pipelines at startup, rather
than lazily as needed.  This CL enables the functionality in Scenic,
woo-hoo!

The goal is to avoid any jank due to lazy pipeline creation in the
middle of a frame.  However, if lazy pipeline creation is attempted
accidentally it will succeed, and a Cobalt event is logged.

Original-Fixed: 44084

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/358888
GitOrigin-RevId: 67b9cdd079bbbc2e549dc071a80f0675b817ce6b
Change-Id: If254392fa56083030cddc26b1f68ca6e1fc7bcf1