[roll] Roll fuchsia [gn sdk]Updating test driver for relative paths
This change is in preparation for runnning the bash script tests
as part of the "normal" CQ. It makes locating the paths of the
scripts to run to be either absolute paths or relative paths where
they are relative to the executable.

Testability: These are tests being updated, so the tests should still
pass, and no new tests that use the relative functionality have been
added yet.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380181
GitOrigin-RevId: aa88f49525e63e1705c76dce45a30c2c5f2d3dd7
Change-Id: Ice2674eb33a614475e1a6a1a9aab0805669a42ba