[roll] Roll fuchsia [fshost] listen for stop on provided lifecycle channel
this takes the lifecycle startup handle provided by component manager
and listens for the stop command. when the stop command is recieved, all
the filesystems are shut down and then the lifecycle channel is closed.

this uses the same shutdown mechanism that is used today to shut down
fshost through the fuchsia.fshost.Admin interface, but exposed through
the new fuchsia.process.lifecycle interface instead. a test is added to
confirm that the lifecycle interface properly calls the existing
shutdown mechanism.

Original-Bug: 45900
Test: fx shell /system/test/ddk/fshost-test-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378731
GitOrigin-RevId: 9768283cdb46e98803a15376492b627e608d4fed
Change-Id: Ie47e4948d39f0496a45f6f70942c64f13ff0e906