[roll] Roll fuchsia [inspect] Validation Archive output
This change modifies the Validator architecture to support validating
JSON returned over the Archive FIDL protocol.

The behavior is triggered by a command line flag, not all puppets are
required to support this feature just yet.

Each operation that occurs on the puppet will be checked both by the
direct connection to the puppet as well as the output from the
Archivist.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/371141
GitOrigin-RevId: 4740a3787510062c30bca164b10c705c22b8860a
Change-Id: Ibcbf93eb20aa579579fb2bce33e83b9abcf2a9b0