[roll] Roll fuchsia [ffx][overnet] Remove SOCAT. Add overnet_channel().
Adds an overnet_pipe() function for comms with a local instance of
overnet with target rx and tx channel endpoints, allowing for
controlling bidirectional comms between two instances of overnet.

This removes the need for SOCAT and the local invocation of `fx onet
host-pipe`, so that one side of the overnet connection is implemented
within the ffx tool itself without need for reliance on an extra fx
command.

Original-Fixed: 46251,49573

Test: e2e tested locally using `ffx run-component`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380254
GitOrigin-RevId: 8afd4cfabd2e69de2bb8c3f93df56f7812fa909c
Change-Id: I8af66924b08c4a722e507f01da254c9bb171b61b