[roll] Roll fuchsia [fidlc] Pass SourceManagers by reference
This fixes a segfault that was occurring in fidlc when
ErrorReporter::PrintReports was called after the relevant SourceManagers
were moved out of scope. The SourceManagers hold unique_ptrs to
SourceFiles, and the Errors in ErrorReporter contain spans which have
raw pointers to those same SourceFiles.

Test: this doesn't lend itself to being tested since the segfault only
happens due to the control flow in compiler/main.cc.
Original-Fixed: 49989
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380313
GitOrigin-RevId: 6239f19f4e726c3e9110b594b90cbd18952bb013
Change-Id: I0f08ece303e67efb16fcecd9011984437b15924a