[roll] Roll fuchsia [ffx][daemon] Add wait for RCS connection.
This adds an implementation to the `Target` class that will allow a
caller to wait for RCS to start given a number of attempts and delays.

This is not an ideal solution, as it would make more sense to use
something like std::sync::CondVar, but async_std and futures don't
support these kinds of synchronization structures.

Original-Fixed: 49659
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379072
GitOrigin-RevId: 29d3e2fe47a3a5bc9dc1522976ddd7e62be98fb7
Change-Id: Id45fa62bb1323f4715e0cc02028fea3bf698a2bb