[roll] Roll fuchsia [retry] Change WithMaxRetries to WithMaxAttempts
The name "WithMaxRetries" left some ambiguity as to the number of times
the operation would be retried. The "max retries" could mean either:
1. how many times the operation is attempted
2. OR how many times the operation is retried after the first attempt
  fails

The retry library actually used approach #2, but based on some of the
usages of WithMaxRetries, it seems that that wasn't obvious to people
using the library.

Additionally, setting maxRetries to zero caused some confusing behavior:
the library interpeted maxRetries==0 as "infinite retries". This means
the only way to turn retries off is to set maxRetries=-1, which is not
documented and is far from intuitive.

I think a decent way to address both of these issues is by renaming the
function to WithMaxAttempts, and changing the parameter to actually mean
"max attempts" (meaning #1) rather than "max retries" (meaning #2). This
also lets us disable retries by setting maxAttempts=1, which is pretty
intuitive. maxAttempts=0 will cause the library to retry infinitely.

To avoid any behavior changes, I incremented the max parameter for all
usages of WithMaxAttempts, except where `maxRetries` was intended to be
be a max number of attempts.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380091
GitOrigin-RevId: e06c5848de6f35d2a9451439b9238d4866440886
Change-Id: I1279a08054caf5074a24201cc5c873c67e4d0467