[roll] Roll fuchsia [appmgr] Fix path traversal sandbox escape.
Since open() was used instead of fdio_open_fd() a path such as
"/pkgfs/../boot/" would be resolved and mapped into the address space at
the destination address. This change moves to fdio_open_fd() to prevent
the traversal of ".." in sandbox manifest paths.

An example of the attack:
"pkgfs": [
        "../boot"
]

Now this resolved path of "pkgfs/../boot" will fail to open since
fdio_open_fd() is being used in replace of open(). This still passes
through all the parent rights as before.

Original-Fixed: 49978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380195
GitOrigin-RevId: 296d77c51130c3e33fd2dc09b49b0b44963193f6
Change-Id: Iec864a28537993fd2e3dd3d791396ea62324193a