[roll] Roll fuchsia [gn sdk]Updating tests to run out of src tree
This moves the sourcing of the library scripts and referencing
test data to be relative to the scripts under test vs. the source
tree.

This allows the tests to be run as part of the "normal" CQ.

Testability: This is a refactor of tests, no functional changes
so everything should work as now with no regressions.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380357
GitOrigin-RevId: f297d0f7e1891d4e7a145ea289ae7df34d47ccd5
Change-Id: Ifca8af938e43ec920355d80282b49ff197ae6088