[roll] Roll fuchsia [tests] audio_core_hardware_tests wait for default
This closes a race condition in audio_core_hardware_tests.
Instead of waiting for an input device to be detected before
beginning capture, the test suite waits until a detected
input device has become the default. audio_core's internal
routing is updated AFTER a device is announced as detected,
but BEFORE it is announced as the new default device, so
this closes a possible race window.

MULTIPLY: `[
  {
    "name": "audio_core_hardware_tests",
    "total_runs": 200
  }
]`

Test: test-code only
Original-Fixed: 43872

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380238
GitOrigin-RevId: d6583009003c2d74db9e4f984683f778c937251c
Change-Id: I6a658355fb51a28e5e4180ba6036e736f461622d