[roll] Roll fuchsia [sl4f] add facade for wlan_phy
The WlanPhy protocol now exposes a GetCountry() method,
to add in testing. Add a facade for the WlanPhy protocol,
to expose the same to ACTS tests.

Design note: we chose to break this out into a separate
facade (rather than integrate into the Wlan facade), because
the WLAN team would like (over the long term) to be able
to validate WLAN at each FIDL layer.

Original-Bug: 47337
Test: manual (see below)

Manual test
-----------
- Cherry-pick https://fuchsia-review.googlesource.com/c/fuchsia/+/368510
$ fx set core.astro --with //src/connectivity/wlan/drivers,//src/connectivity/wlan/testing/wlantap-driver:wlantap,//garnet/bin/sl4f && fx build && fx reboot -r
- On an AOSP checkout, cherry-pick https://android-review.googlesource.com/c/platform/tools/test/connectivity/+/1246540
$ act.py -c ~/configs/fuchsia_testbed.config -tc WlanFacadeTest

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/368511
GitOrigin-RevId: 84316e5259aff3b9766d914b9c99b96a6fbe2126
Change-Id: I910b40cd8eec490358dd1a52d0105fba40633d31