[roll] Roll fuchsia [trace_processing] Add frame latency to flutter_frame_stats
Adds a metric that measures the time between flutter starting to produce
a frame and it getting displayed on the device. This is intended to
provide an alternative to input latency metrics for use cases where the
output is driven by a programmatic state change rather than a user
input.

Original-Bug: 49925

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375002
GitOrigin-RevId: 0e6d2f2c0248343202a33fb6c0e43a72ced8e841
Change-Id: Iaea436eab270b8274a586edc1b0f8a57eb5550a7