[roll] Roll external/github.com/catapult-project/catapult Dashboard: Reduce estimator function overhead
This change reduces the cost of the estimator function by half using the
squared differences of combinations of values in the intra-cluster
computation. We also change this so that we only run the estimator
functions when necessary:

- Once when we're finding a potential partition point for a range.

- When we're determining a potential change-point for each permutation
of a sub-range.

This should suffice in the meantime until we can use a more efficient
numerical computation library (like numpy, part of scipy) which
leverages SIMD computations to perform the heavy math parts of the code.

Local testing shows an 80% reduction in runtime, with improved noise
reduction. We've had to adjust the contrived test cases to ensure that
we're finding the change-points we are expecting to see in the synthetic
cases. No changes were necessary to the tests which include data fitting
the kinds we have encountered in the wild.

In this change we also roll back the creation of `array.array` instances
because those turn out to be expensive in the inner loops.

R=abennetts@google.com, fancl@chromium.org

Original-Bug: chromium:1070268
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2148558
Auto-Submit: Dean Berris <dberris@chromium.org>
GitOrigin-RevId: e1f6e1133a846cb2ef8b2beaf925dc9bcd2602e7
Change-Id: Ib540c80e987c4fc74759f7d1f7a84fd93379e465