[roll] Roll fuchsia [brcmfmac] Fix inconsistent scan issue for association
Changed to non-scan association so that there will not be a active scan
being sent out before join. Because non-scan association has a
unreliable issue before the association retry being added, and that
change is rolled into the tree now, so we can safely have this change.

Non-scan association support is added in SIM, ssid field should be
added in SimAssocReqFrame as well, for fake APs to verify whether it's
a correct request.

Original-Bug: 43096
Test: wlan-stress-test -c fx-apple -r 10 assoc_test(modified),
assoc_unittests, tests under sim-fake-ap/test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/365715
GitOrigin-RevId: 5097a14430821060f0dc8d5d2cf33db456bb5fe6
Change-Id: I26095c529f39a46424daa353ce4b9c387cf2ea3a