[roll] Roll fuchsia [kernel][vm] move a definition of a global array of kernel regions out of a header
Move the body of the array definition into a .cc file to move its
logic out of a header file and prevent multiple definitions of the
same array.

No functional change.

Tested: build and boot on arm and x86 qemu, validate that the array is
properly iterated and used in boot up and unit tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380196
GitOrigin-RevId: e9faac2ce2edeb439e4ce43940935a582a0b9d40
Change-Id: Id6b064f82a18e0fb3be4bb679c9968b8b260645e