[roll] Roll fuchsia [kernel][x86] Ensure that turbo state matches command line state.
Currently, we explicitly disable turbo on each CPU when the kernel
command line argument disables it, but won't actively enable it.
This means that if previous kernels in the boot chain have disabled
it, we will continue to run without Turbo enabled.

This CL will unconditionally enable/disable Turbo based on the kernel
command line arguments.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379543
GitOrigin-RevId: e0a1ea682d1313582f8eb7482dbe8a5db4d1eb31
Change-Id: Ieaf163b5ce827b6b4504d6eb03f8f5ef8c4c10e4