[roll] Roll fuchsia [chunked-compress] Introduce chunked compressor library
This change introduces "chunked-compress", a library for creating
compressed archives that support decompressing sub-segments of. This is
intended to be used to support demand paging of compressed files, for
example in Blobfs, without having to decompress the entire file.

The format is fairly simple: A header, a seek table, and a number of
frames. Each frame is a regular ZSTD archive.

Detailed documentation can be found in "archive-header.h".

This change only adds the compressor (decompressor is in a child CL).

Test: Extensive unit/manual testing.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378774
GitOrigin-RevId: cc7ed466d482a3e112ce0902fef6470b5b56a683
Change-Id: I75bbd8535146b1937a2ca6adc5db8a570d4a16e7