[roll] Roll fuchsia [bt][audio-output] Drop audio socket with stream
When the audio stream was dropped, the SoftPcmOutput thread
handling the audio device would continue on, with the audio system
unawares that no one was listening.  Add a handle so when the audio
output stream is dropped, the background task is also stopped.

This drops the audio device channel and any RingBuffer channel
established as well.

Original-Fixed: 43812
Test: added soft_pcm_audio_should_end_when_stream_dropped
      fx test fuchsia-audio-device-output-tests
      manually tested with bt-a2dp-source output
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380436
GitOrigin-RevId: 4beae6c0217dc5c55d9735747d10c41167104cdc
Change-Id: If7d81e636a49352a55df658492f5444abf3ae7a3