[roll] Roll fuchsia [blobfs] Return error if syncing a blob too early.
Syncing a blob to disk before its data is complete has no meaning
because blobs are content-addressed. Without all the data, the name can
have no meaning.

Previously calling Sync() on a blob in blobfs would immediately return
ZX_OK. But this is misleading because nothing was synced. This change
makes that case return ZX_ERR_BAD_STATE.

Move the code from Blob::PopulateInode to a new constructor and have
callers use that instead of manually populating the Blob. This cleans up
the callers and makes state management a little more clear. SetState()
and GetState() are now be made private.

Clarify threading behaviors in Blobfs and Minfs.

Removes a thread from the blob-test.cc harness. Blobfs wants a
dispatcher for the current thread, not a new background thread as was
being used before.

Original-Bug: 33099, 49831

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379379
GitOrigin-RevId: 80aa98a45d73b16d87c152d774c888d3bce352d9
Change-Id: Ib5d19ed4bf8a6abffbb8787605720bc6b5caffd3