[roll] Roll fuchsia [validator] Match Inspect data against Archive output
This change enhances the Validator to actually check that the local data
being validated matches the output of the Archive. This required several
changes to validation:

- Formatting consistency for all to_string methods. Depending on node
nesting we could end up with several newlines at odd places. This change
joins strings together to avoid newline proliferation, and it removes
internal spaces.
- Diff consistency. try_compare now uses the difference crate to compare
the to_string outputs, and then prints them in an easy to read format.
- Consistent output across all data types. The Archivist follows and
combines data from link nodes, which was difficult to match against in
the Validator. This change ensures that the contents of link nodes are
exposes consistently such that a LazyNode/LazyValue link will give
identical output to a real hierarchy with the same nodes.
- Generic data types for parsed JSON. JSON cannot represent every type
supported by Inspect natively, so this change introduces the concepts of
GenericNumber, GenericArray, and GenericHistogram. We provide conversion
for local Data and parsed Data to these types so they can be directly
compared.
- Omission of duplicate names in generic output. When you use to_generic
(for JSON comparison) any duplicate properties will be ommitted. This is
to support cases where we test how this appears in the binary output
without needing to disable Archive tests (duplicate keys cannot appear
in JSON).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379239
GitOrigin-RevId: 2ef8f44a1f56fd44cbf0597a13bd8a781c70cefb
Change-Id: Ibba5314cdf644c4aa9a4462d9bcb91f97b27c3a9