[roll] Roll fuchsia [swd-e2e-tests] Consolidate test paving logic
This extends `device.Client to take advantage of `Build` by adding a
`device.Client.Pave` method that paves the device to that specific
build.

Furthermore, this extracts the test paving logic into
`system-tests/pave`. This module supports both rebooting to recovery, as
well as OTA-ing to recovery, so this patch also adds support for this to
the reboot and tracking tests.

Test: Verified OTA tests still pass with this CL applied.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378869
GitOrigin-RevId: e66f534396c42406246b8f309989e33ab934dc79
Change-Id: I607198d1cf8d2bb31c49a6d7f3dfbc4083965d33