[roll] Roll fuchsia [bt][a2dp-source] Respond to peer requests
Connect MediaTask Streams to the request parsing of Peer, meaning it
will respond, configure, start, suspend, etc. streams based on peer
initiation.

This is required for many remotely-connecting sinks to connect and
stream audio, such as when many speakers are power cycled and reconnect
to their last connected device.

Original-Fixed: 2783
Test: Manually tested with headphones
      Added tests in a2dp-source::peer::Peer
      fx test bt-a2dp-source-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377980
GitOrigin-RevId: cabb776fcebcab06a49abed4e6739275de20c1ce
Change-Id: Ie85b17eb2526442787f7e415f039b9f5acb4ef9d