[roll] Roll fuchsia [audio_core] Fix DeviceDefaultChanged callbacks
The state machine for device plug-changes and removals is adjusted to
preserve the invariant of "notification that a device is no longer the
default must always precede the notification that it is removed."

Specifically, the device removal process always routes through the
device-unplug state, relying on OnPlugStateChanged always calling
UpdateDefaultDevice.

Related, AudioDevice::UpdatePlugState previously would not change the
plug state of a device if called during the device's wind-ddown
process, but this was preventing removed devices from triggering
OnPlugStateChanged's call to UpdateDefaultDevice, thus preventing
the default-changed callback from notifying clients. Depending on
timing, and if a test was relying on the 'default' mechanism to know
when media was being routed to the device, this could cause
inconsistent test results.

Test: New case in audio_device_tests that fails without the fix

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379835
GitOrigin-RevId: 5f988c959e0f06792a78c76074ad3b1693adc690
Change-Id: Iadb5ded6dd32769abe0c1da0dd8f461db2197f7f