[roll] Roll fuchsia [cobalt] Fix flake in testapp
Original-Bug: 49548
Original-Bug: 50066

cobalt_testapp_no_network has been flaky and looking through the logs,
the failure is almost always TestDebugMetric. This turns out to be an
obsolete test anyway as it is was designed for testing a now obsolete
mechanism for determining when to enable debug metrics. Therefore I
am just deleting the test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380456
GitOrigin-RevId: 893f078fe9aee405a420bdeba901b5b1679da967
Change-Id: Ib2272f4c206ade34908e22340cb24414c9864d6a