[roll] Roll fuchsia [audio_core] Account for reported external delay.
When sampling from renderers, be sure to account for the external
delay reported by the driver when computing where to sample from.  For
built in speakers and such, this tends to be a small amount (100-150
uSec at most), but when we are using devices like bluetooth
headphones, the numbers become rather significant (100+ mSec) and need
to be taken in consideration in order to achieve anything like good
sync.

In addition, start the process of cleaning up some of the names used
for things here.  We are moving in the direction of allowing inputs
and outputs to use clocks which are not clock monotonic as their
reference clock.  Stop using the term "clock mono" when we mean to say
"ref clock".  Also clean up some of the names of the transformations
exposed by the AudioDriver.  Describing a mapping from the reference clock to
frames is not precise enough, because it does not say _where_ in
the frame sequence we are mapping the reference time to.

There are three main positions in the frame we usually care about.
They are

++ The point of presentation or capture.  In other words, when does
   this frame hit the speaker or get produced by the microphone.
++ The point of transmission to or reception from the interconnect.
   This is the point at which the audio data either enters or exits the
   chip.
++ The point of the "safe" write or read pointer.  This is the point
   in the frame sequence where it should be safe to either write a new
   frame for presentation, or to read a frame which had been captured.
   Getting anywhere closer to "now" from this point runs the risk that
   the hardware may have either already read the data from memory into
   its transmission FIFO, or may have not yet moved the data from its
   reception FIFO into memory.

Produce transformations for each of these which should be valid as
long as the driver has not been stopped and stash them at the
AudioDriver level to be used as needed.

Test: External verification along with a new unit test which sanity checks the
      math done by AudioDriver.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378559
GitOrigin-RevId: b5b175dd21378dc72ad4a5aca3909cd34bd9d9c9
Change-Id: I5d9fe8f3a2cf0a62e8bab24156fec441fbe23ce8