[roll] Roll fuchsia [benchmarks] Fix blobfs-bench-test invocation on garnet-perf builder
This invocation was broken by this change:
"[unification] Move tests from manifests to packages"
https://fuchsia-review.googlesource.com/c/fuchsia/+/376514

That wasn't caught because the garnet-perf builder (which now only
runs storage performance tests) is not run by the CQ by default.

blobfs-bench-test was moved from the Zircon GN build to the main
Fuchsia GN build.  Referring to it via the zircon_extras_manifest hack
no longer works; it's now in a normal Fuchsia package.

Original-Bug: 49874
Test: garnet-x64-perf-dawson_canyon trybot
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379873
GitOrigin-RevId: 624e99dfbc13e4ce7e381db9ad839af9d2c62c2e
Change-Id: I8acbd1fc2941f7185e59090d47d640aa7cd23c84