[roll] Roll fuchsia [logs] Type-safe severity parsing.
In order to support a new severity scale, we need to support two
sets of ordinals in the wire format simultaneously. All of the
meaningful *readers* of this value are in-tree, so we plan to do
the migration by rewriting old->new in the Archivist.

To be able to rewrite the current/old severity values into the new
severity values, we need to parse the current values (which we
previously didn't do at all).

Original-Bug: 49474
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378973
GitOrigin-RevId: f115656334a0499790163b41d46ccd27f6a8d6db
Change-Id: I0cac3426481180c58239a9be0c419f2a77a1e6f4