[roll] Roll fuchsia [amlogic_encoder] Use OutputSink to implement buffer handling
Use the helper class OutputSink to handle output buffers in the
encoder codec adapter since it handles synchronization of buffers
coming and going for us. Fill in a dummy encode frame implementation
and update the integration test to expect an output packet.
This validates that codec adapter buffer and format negotiation completes.

Test: fx test h264_encoder_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/369280
GitOrigin-RevId: fcf50349c4b4407368b3411b2108d5ac880af9c8
Change-Id: I4f09270973e437e8eac9e20a613fcddde95a35b2