[roll] Roll fuchsia [GN SDK] Moving data dir to $HOME/.fuchsia
To avoid having persistent settings and SSH keys inadvertently being deleted,
this moves the data directory used by the SDK scripts to be in the
user's HOME directory. The default location is $HOME/.fuchsia. For
testing or general personal preference, this can be overridden by
setting the environment variable FUCHSIA_SDK_DATA_DIR.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377145
GitOrigin-RevId: 738d5a8cbae4c8b78a43089020048ab0d1d716a7
Change-Id: Ic9725b44ed96e14cf2e765adfb54ac003cfb717d