[roll] Roll fuchsia [swd-e2e-tests] Use ctx in sl4f; wrap major phases in timeouts
Previously, the e2e tests used contexts to provide a timeout mechanism,
but unfortunately this relies on contexts being used in every
potentially long running call. We missed one in sl4f when we make an
http request. This hung in a test, and resulted in the whole test
hanging.

This addresses this hang in two ways. First, we pass the context to the
sl4f http client, so if the context gets canceled, the http request
should also get canceled.

Second, we extract out the pattern of "run a cancelable function in a
goroutine, and wait for it to finish, or for the context to cancel. We
then use this functionality to time out the major phases of each test in
case they hang. Finally, we migrate all the manual implementations of
this pattern to use this function to reduce code duplication.

Test: Ran tests, verified they still pass locally
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380762
GitOrigin-RevId: 6067fc4f3605030f5d9f2ba6c81f0c0f02eca852
Change-Id: I482dff7a72ee171860fe6bd23a38a41e7e693aca