[roll] Roll fuchsia [drm] Add bounds to vectors and strings
The FIDL API rubric recommends putting bounds on all vectors and
strings and the fidl lint tool enforces this. This CL adds bounds to all
vectors and strings in the fuchsia.media.drm library and removes the
fidl lint exception. All bounds chosen are well within current limits
and no uses need to be transitioned.

Original-Bug: 35879
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380694
GitOrigin-RevId: cfab8022a032ffdb39b88b4aef1e6d64828447e8
Change-Id: I97b0a9b34e45efac536226969f2bf9d62f853b90