[roll] Roll fuchsia [debugger] Try to fix racy test.
Try to fix a flake in the debugger-test. The problem was that waiting
for the SUSPENDED notification would occationally result in a thread
that was RUNNING. The best theory is that the thread notifications are
racing from the test's previous unsuspend of the thread.

This change forces the target thread to be in a RUNNING state before
waiting for it to be SUSPENDED again.

Original-Bug: 48822

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380166
GitOrigin-RevId: 625465c8d26eb94777260ccb84fa56e39d8286f9
Change-Id: Ia151db56bba3203834b21b874c8f69eb8e86101c