[roll] Roll fuchsia [recovery] Make recovery UI a Carnelian app
Make the recovery UI as full Carnelian app,
instead of a mix of direct use of the frame
buffer and Canvas. This is needed for two
reasons; first is that canvas will be going away
in favor of the render interface, second is that
the recovery UI needs to be rotated on some
hardware and the render interface makes that much
easier.

This port should not be a performance regression
from the canvas-based code, but it could be
greatly improved by not redrawing the entire
screen on each vsync.

Testing: ran on NUC and tested hitting the HTTP
server with curl, ran unit tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378565
GitOrigin-RevId: 2eca1904b20e5203c9c39197f4b844e324610b75
Change-Id: Ied9975f36d8cc1e9afcffc7addcb5c3c0139809d