[roll] Roll fuchsia [flatland] Move allocation responsibility from server to client.
Previous version of BufferCollectionInfo.h/cc had the server side
responsible for waiting for allocation but this CL moves that
responsibility onto the client, and updates the BufferCollectionInfo
class as necessary to make that happen. This is important because with
this new change, the server no longer contains any blocking code.

Concretely, BufferCollection now has a "BuffersAreAllocated()" function
instead of a "WaitForAllocatedFunction" that immediately returns
false if the collection is not allocated, instead of blocking.
Additionally, the GenerateToken() function, as well as the constraint
token, are removed entirely from BufferCollectionInfo and it is now the
client's responsibility to duplicate whatever extra tokens they need
before moving the token into New().

Testing: Existing unit tests in buffer_collection_unittest.cc and
gpu_mem_unittest.cc have been updated to reflect these changes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380393
GitOrigin-RevId: 98f3cea4531cb0a8b739a30bdf9d354c593c0e7d
Change-Id: I7f42293bb9a603c7bad305bff5b07302e714816e