[roll] Roll fuchsia [fx][create] Add golden project testing.
Golden projects checked into source control are tested as a normal
part of the build (//tools/create/goldens/*).

To test that the generated project files match the golden project
files, a new target rule `golden_test` is introduced.

This rule defines the arguments to pass to fx create and the set
of golden files to compare the output against.

These tests are run with `fx test //tools/create/golden:tests`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376806
GitOrigin-RevId: 668e5a9412338460c9b729e5d6a4bca10c8923dd
Change-Id: I8deafe69e4f7500adbaa4e3c1263d132f60180a5