[roll] Roll fuchsia [debugger] Fix IPC version check.
We were never actually checking the IPC version which could lead to
surprising effects if the versions mismatch. The cause was that there
were two constants, the one being used by the backend, and the one we
were actually updating when we changed the protocol.

This change removes the one we weren't updating.

The message wording is updated to make it more clear and to force
terminate the connection to reduce confusion.

Test: manual (tests wouldn't have caught this since we were changing the
wrong place when updating the protocol).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380986
GitOrigin-RevId: 6aa7ac35a463c990623b377dd680c7173352f5c2
Change-Id: Ibb3896cacd0369e3dc14f32806ff57f563256dae