[roll] Roll fuchsia [ddk] ShutdownClientConnections after unbind finishes
This patch forcefully shuts down client connections after a driver
replies to Unbind. This is needed to ensure that DdkRelease can
eventually be called if a client does not close its channel.

Additionally; we now wait for a ZX_CHANNEL_PEER_CLOSED
signal even when we are unbound. This is nececssary
to prevent a race where we get a message followed by
an immediate unbind, which previously would result
in the handler being de-registered.

Test: runtests -t ddk-lifecycle-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380300
GitOrigin-RevId: 88e4d08a6cd5aef5fd0d17891867b66f181b5859
Change-Id: I0f08db5d7d8f5044aaebdb8359d8ca9c15cc7984