[roll] Roll fuchsia [SetUI][core] Add Message Signature.
This changelist introduces a signature to a message.
This identifier can now be used as an audience for a
message, allowing those who receive messages from
brokers or other unaddressable sources to send top
level messages back. It also allows for brokers to
contact the author of messages.

Signature is part of a messenger's Fingerprint, which
also includes its MessengerId. This structure is now
part of Messenger and passed with requests to the
MessageHub.

This change also includes fixes to behavior when
the broker is the author of messages.

Original-Fixed: 50211
Test: A new test has been added that verifies top level
messages can be sent to a Signature captured from an
earlier message.
fx test -o setui_service_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380954
GitOrigin-RevId: c1ba6bab856210087b9e643d5546220f146327ae
Change-Id: I1fe17f5fdbcb858ebbd0941c5b2a56030ebe97a5