[roll] Roll fuchsia [amlogic_decoder][h264] Allow inputting multiple nal units at once
The hardware needs the raw SPS and PPS data, so the existing code passes
through an entire buffer to the hardware if it detects that it begins
with an SPS or PPS NAL unit. That doesn't work if there are multiple NAL
units in the input, so modify the chromium code to store the raw data
when it parses an SPS or PPS. Then that data can be passed directly into
the stream buffer before decoding a frame.

Original-Fixed: 50146

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381171
GitOrigin-RevId: ea0888e7a5347126659a7395132f15ab9d17f5ad
Change-Id: Ic24b6fc65c0018ca49f6114d4854276951c48704