[roll] Roll fuchsia [chunked-compress] Add optional per-frame checksums
Adds a configuration knob to CompressionParams to enable including a
per-frame checksum. This is internally checked by zstd during
decompression of each frame.

This is not a breaking format change, as the inclusion of a checksum in
a frame does not require different behaviors during parsing or
decompression.

Tested: Unit tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379493
GitOrigin-RevId: f7dbaca5def21c08d780de0b193fe8b150cde173
Change-Id: I23f5c7a197176deecb0bf76bc0582304d63b6dcb