[roll] Roll fuchsia [components] Introduce EventResult to track state transition errors
This CL introduces an EventResult within an event object.

An EventResult is defined as a Result<EventPayload, EventError>.

When a state transition starts in copmonent manager, a bunch of work
happens. This CL bundles the error with the pending state transition to
give more context on the error.

This error can also be propagated via FIDL to be used by Archivist.

This is the first of many CLs that improves error reporting in component
manager. This CL collects component resolution errors and propagates
them out.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378413
GitOrigin-RevId: 81caa801cc33f7f4f81189ac0c98950a03ff0e7b
Change-Id: If5786a7c9ca5e8b100d8e1c8e03c23ac62bfacb4