[roll] Roll fuchsia [kernel] Add RelaxedAtomic helper.
Wrapper for ktl::atomic that exposes a limited subset of methods that
default operations to ktl::memory_order relaxed.

The goal is mostly to eliminate ktl::memory_order_relaxed noise from all
of the call sites. But it also forces all uses of the atomic to be
relaxed so you don't have mixed interation types.

Test: Added unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/360065
GitOrigin-RevId: 691a16635362ab82a0258ff347e7dc63ac159888
Change-Id: I3a2c5aaef9e146fc5b56fe44a396114ea82d77ec