[roll] Roll fuchsia [testrunner] Distinguish output files for multiplier tests.
Test results are stored by test path or package url. This causes the
output of multiplier tests (the same test run multiple times) to be
written to the same file, so we append an index to the test name so the
logs get written to separate files and get reported correctly.

Original-Bug: 46142
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381218
GitOrigin-RevId: f7293e35dabdb15ecad76a2b61adabd9ba53d2f0
Change-Id: I5ed3599ace3e6443a74222e71a9daf09f1192217