[roll] Roll fuchsia [microbenchmarks] Reduce chance of hitting Infra's 5-minute IO timeout
Swarming tasks on Infra are aborted and marked as "timed out" if they
run for a certain amount of time without producing any log output.
For the terminal-x64-release builders, this amount is currently set as
"swarming_io_timeout_secs: 300", i.e. 5 minutes.  (This parameter
shows up as "I/O timeout: 5m" on the task pages on
https://chromium-swarm.appspot.com/.)

The first invocation of fuchsia_microbenchmarks is currently only
taking about 1.5 minutes on NUCs.  However, changing Intel HWP
settings (CPU frequency scaling) to "power saving" mode appears to
push up the time to around 5 minutes (with microbenchmarks_test.dart
taking ~7 minutes overall), which would make the test invocation
flaky.

To avoid that problem in the future, halve the number of
within-process iterations of each test case from 1000 (the current
default) to 500.  While we could increase swarming_io_timeout_secs, we
don't really want to be spending so much time running
fuchsia_microbenchmarks.

There are diminishing returns from running within-process iterations
anyway: Doing 1000 within-process iterations is probably excessive,
and it is probably better to spend the resources on more process runs.
This can be quantified, but the tools for doing that aren't
well-developed yet, so I'm not trying to do that for this change.

Original-Bug: 49280
Test: microbenchmarks_test on terminal-x64-release
 * tested locally with: fx run-e2e-tests microbenchmarks_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380968
GitOrigin-RevId: a72c8be02cb727751ec01ab04654a7e944239ebb
Change-Id: I664d60335172d45c2d55452226b348d260a89278