[roll] Roll fuchsia [bt][gatt] Client: Read by Type Request
Add support for the ATT Read By Type Request to the GATT Client. This
allows the client to obtain the values of attributes by UUID when the
handle is not known.

This feature is needed for PTS test GATT/CL/GAR/BV-03-C.

Original-Bug: 48293
Test: in bt-host-unittests
GATT_ClientTest.ReadByTypeRequestSuccess
GATT_ClientTest.ReadByTypeRequestError
GATT_ClientTest.ReadByTypeRequestInvalidResponses

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378540
GitOrigin-RevId: 6b645d73541260bf7d762ff88d52f91a72cfd815
Change-Id: I0ee4c65aa8024a84e6db3265ff9589c03e997f18