[roll] Roll fuchsia [component_manager][events] Add a timestamp to events
The timestamp is added at the hooks::Event level
so that the same timestamp that is sent on started can be re-used
when synthesizing Running events. This also gives a more accurate
timestamp of when the event originally occurred rather than when it
was dispatched.

Tested: component_manager_tests, events_integration_test
Original-Fixed: 50261

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381158
GitOrigin-RevId: d46ec70cfd3a30fb63436452950bf240cceb6e3f
Change-Id: I0bca6528f1293c635f27015588ba8c7b09f3a63d