[roll] Roll fuchsia [brcmfmac] Move WLAN component/fragment enum to wifi-config.h
The component enum values define the order of the individual devices in
the composite WLAN device, so this ordering should be defined in the
common wifi-config.h header.

In addition, the "components" of a composite device have been renamed as
"fragments" in the DDK, so we make the corresponding renames as well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/370694
GitOrigin-RevId: fe513fdc071786cd8a81958f49b0fe16ffdda9f0
Change-Id: I14b5a148349f6105e46aff90a080e83689d4c0e2