[roll] Roll fuchsia [fidl] round-trip benchmarks for FIDL bindings
This runs round-trip tests of a few different FIDL messages. The runner
is written in Rust and there are peers for LLCPP and Go. Other peers
will be added in subsequent CLs.

This CL is part of the way towards integration with the performance /
benchmark dashboard and alerting infrastructure, hence the enormous JSON
output.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/359213
GitOrigin-RevId: 99dc631713df2e6f30b25deaaeb9ccde9e0b2782
Change-Id: If172f76131e308488334ce2999560e6bd758a618