[roll] Roll fuchsia [fidlc] Format Errors in ctor
An Error<Args...> currently stores a std::tuple<Args...> of its format
parameters. This allows for introspection of those parameters after
error creation/reporting, and allows for deferring formatting. In
practice, this isn't actually used, and causes problems because many
Errors' parameters have dangling pointers to AST nodes that may have
been freed by the time ErrorReporter::PrintReports is called and the
Error needs to format its parameters.

This CL replaces BaseError::Format with a pre-formatted std::string,
which is set in the Error constructor.

Test: fidlc host tests
Original-Fixed: 50262
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381389
GitOrigin-RevId: 3d3ddb07b7b73b51157fce46c2740559aa664573
Change-Id: I451273aa5b3babfe7af0bd4b0f2c26cc4d47ee8c