[roll] Roll fuchsia [build][fuzzing] Add seed corpus support
This CL adds the GN logic for fuzzers to include a seed corpus
directory. These should contain hand-crafted examples to increase fuzzer
coverage and artifacts from previous fuzzing runs to test for
regressions. These elements will be include as resources in the package,
to be available either for fuzzing bootstrapping, or unit/regression
testing when not fuzzing.

Original-Bug: 49418
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377753
GitOrigin-RevId: bf8623cc8e3333d18fc0392c3d65bc91e43d36e3
Change-Id: Iacccbc08f625037d0a1ee3e284f39aed05db2e34