[roll] Roll fuchsia [archivist] Prepare component id management for component v2 monikers
Currently the archivist managers component ids for v1 as a triple:
(component name, component id, realm path).

In v2, no such notion exists and instead we only have a single thing: moniker.

For this, ComponentIdentifier is introduced to hold both v1 and v2.
Some places where the triple was used and now we'll need to use a single
value, have been converted to realm_path/component_name:instance_id as
in components v1 names are not unique so the instance id is needed to
disambiguate.

Tested: archivist_tests, archivist_integration_tests
Original-Bug: 50226

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380560
GitOrigin-RevId: b6ca898b9dc11015b62349d5ab1f562d3051a135
Change-Id: I5c84206fb22ca26ed714a384a283256b7390e76d