[roll] Roll fuchsia [bootserver] add firmware support
Adds support for sending firmware to the device.

Firmware is a superset of the old "bootloader" file type. Instead of a
single file, it supports any number of firmware files by including an
optional string type.

These types are device specific, so e.g. if a device has multiple
bootloader stages it might support firmware "firmware_stage1" and
"firmware_stage2".

Critically, failure to write an unsupported firmware type is not a
fatal error, but is just skipped. This is important so that we can add
more firmware to an OTA package later, and older devices will safely
ignore it until Zircon has been updated to the point where it can handle
it properly.

Test: `bootserver_new --firmware-unknown <file>` properly ignores the
      unknown firmware type (tested on Astro)
Test: `bootserver_new --firmware <file>` properly writes the known
      firmware type (tested on Astro)
Original-Bug: 45606
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380984
GitOrigin-RevId: b076e21fb88f754d33ddf1e91c2516ef57e6275f
Change-Id: Ia606c0d1104e7b5253774f7f9eda9f99bb7f3b3c