[roll] Roll fuchsia [memorypressure] Make accesses to watcher state more robust
1. Use |std::vector<std::unique_ptr<WatcherState>>| to store all
registered watchers instead of |std::vector<WatcherState>|.  The
|WatcherState| was previously passed by reference to lambdas which
handle sending out notifications, and callbacks for these notifications.
Passing around a reference to the vector element is unsafe if the vector
is modified and the backing memory location changes. Use a vector of
|std::unique_ptr<WatcherState>| instead and capture raw pointers in
lambdas.

2. The |std::unique_ptr<WatcherState>| is removed from the |watchers_|
vector in case of an error on the watcher channel.  If an error occurs
while a callback is pending, delay freeing the |WatcherState| until the
callback is executed. The callback captures a raw pointer to the
|WatcherState|, so removing it from the vector before the callback is
executed can lead to a use-after-free.  This ensures that the provider
is robust in the face of races or crashes in the watcher, where the
error handler is not the last operation on the |WatcherState|.

Test: Added test cases to PressureFidlUnitTest. Also verified that the
Chromium crashes reported in the bug no longer occur.

Original-Bug: 48322
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380495
GitOrigin-RevId: 5e798d5d4e9a84b7df3d8b175759d7a3eb11cd04
Change-Id: I9ebbf5a6e830a804aedd277c73a3f60c72f633e9