[roll] Roll fuchsia [chunked-compress] Cap headers to 32KiB (1023 frames)
Clients need to know some sort of reasonable upper bound for how big a
header could be to read-ahead the seek table.

Arbitrarily select 32KiB, since this is 4 data blocks in Blobfs which is
a reasonably small amount of data to prefetch in order to parse the seek
table.

Tested: Unit Tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381735
GitOrigin-RevId: d813457778bf2f0b127de6a084291c11a518d2cd
Change-Id: I13606c79959939c72681857c6f86d518369f6c8d