[roll] Roll fuchsia [pci] Avoid triggering move constructors.
Take advantage of std::make_unique's perfect forwarding behavior and
construct elements directly instead of constructing an element on the
stack and then rvalue constructing the heap allocated object from the
temp object.

Generally, stuff like this should not matter much as the compiler will
optimize the generated code.  This does technically trigger the move
constructor, however.  We are in the process of locking down the
moveability/copy-ability of elements contained in intrusive
containers, and triggering this move constructor would force the code
to opt-in to allowing rvalue construction of Capability and
ExtCapability, which currently there is no reason to do.

Tests: No functional change.  Exiting tests only.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381336
GitOrigin-RevId: fd009db93f618f7fc9617585d02b07d015051ce3
Change-Id: I540df36853846563747437bd52f6326107a7a781