[roll] Roll fuchsia [storage][volume_image]: Extent util.
It is phrased as extent so it is not limited to extent -> slice
conversion, and can be used across our tooling. Meaning we have a single
way for doing this.

Data that is in contigous blocks, would get repartition if the blocks
are not multiple of each other, but that would fit in the virtual
address of the vpartitions. Otherwise we would need to account for the
tail.

A tail is added so we can figure out where and how much we need to
fill(e.g. zero intialize) the remainded of the last block of an extent.

TEST=storage-volume-image-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380455
GitOrigin-RevId: 108fffc226db6144c82be65f204927cf0c6a4d81
Change-Id: Id7aeecf09ed190c1837a251be2ab0bb4f60a0d29