[roll] Roll fuchsia [sshutil] Client shouldn't use SetDeadline
sshutil.Client originally used SetDeadline in it's keepalive loop in
order to disconnect when we didn't hear back from the server. Since
then, we added support for a timeout channel, so the SetDeadline is
redundant.

In addition, this might also be causing a deadlock in ssh library, as
it's internal channels might not be getting closed if the underlying
connection is getting closed by the deadline handler. I've been unable
to prove this yet so far though.

Original-Bug: 48756

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381179
GitOrigin-RevId: 2d1649f901e7779260499784568da41ce52f8f8f
Change-Id: I58a3c1b3b7bffb7e9d85dc2048c0cafcb300097d