[roll] Roll fuchsia [fbl][containers] Add options to fbl intrusive container nodes.
Add the ability to pass a set of flag-style options to intrusive
container nodes via template arguments.  This will allow us to control
and sanity-check various fine-grained behaviors of container nodes at
compile time.  For example, we can use flags such as this to disallow
copy assignment of any node at compile time unless the user
specifically opts into the behavior by customizing their node with a
flag.

The majority of this CL, however, deals with some restructuring of how
ContainableBaseClasses works in order to make it easy to not only pass
flags, but also use the tagging extensions to easily put objects into
multiple containers at the same time.

Prior to this, ContainableBaseClasses used inheritance to allow a T to
exist in multiple containers at once.  Effectively, T written as

class T : public Lisable<T*>,
          public ContainableBaseClasses<
                   Lisable<T*, Tag1>,
                   Lisable<T*, Tag2>,
                 > { ... }

Would expand to a class hierarchy which looked like

+--------------------+
| Listable<T*, Tag2> |
+--------------------+
        |
        v
+--------------------+
| Listable<T*, Tag1> |
+--------------------+
        |
        v
+--------------------+   +---------------------------+
|  Containable<...>  |   |  Listable<T*, DefaultTag> |
+--------------------+   +---------------------------+
        |                        |
        v                        v
      +----------------------------+
      |             T              +
      +----------------------------+

When default traits were used to find a given node, they used the
specific _type_ of the node to find the node in the hierarchy by
effectively saying:

  using NodeType = Listable<PtrType, TagIAmLookingFor>;
  return obj.NodeType::node_storage_;

Adding templated options to this mix makes this using this technique
more painful.  Even though every type involved in the hierarchy is
uniquely identifiable using the TagType, the code would _still_ need
to know the specific options for a given node in order to know the
specific type for selection in the hierarchy.  While it is possible to
do this, it forces the user to specify redundant information all over
the place, and generally makes the API unpleasant.

So: the major change here is to ContainableBaseClasses.  Instead of
using inheritance to do its job, it now uses encapsulation instead.
The instances of the nodes are stored in CBC directly in a tuple, and
metaprogramming tricks are used to find these instance offsets at
compile time.  This allows our default traits to not need to know
anything about the underlying node aside from it's tag.  The structure
of CBC already ensures that the tags themselves are unique.

With this change in place, we can now extend the node types themselves
without needing to teach the default container traits about each of
these extensions.

Tests: New unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380494
GitOrigin-RevId: ff51a7e6b365a503ae62522a7d552c15da500b4b
Change-Id: Id69994969ca2b90baba8fba59c23942bff53cf9a