[roll] Roll fuchsia [logging] Modify varargs handling in macros
Modify the way varargs are handled in the FX_LOGF and FX_VLOGF macros so
that they can be used by other macros that pass along additional
information in specific contexts.

The goal here is to remove all of the custom logging from the driver
framework, and to rely on syslog.

Original-Bug: 49457
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381048
GitOrigin-RevId: 577957ffcbd04a263a7cec44c1e4ec551f360e97
Change-Id: I81716f1d769f223004020e92750be40b05abb4e9