[roll] Roll fuchsia [gn sdk]Refactor to generic shellscript test
This refactors the far_analysis.gni template into
shellscript_test.gni which is intended to enable creating
bash based host tests to inspect the output of the test build.

This is in preparation for adding tests for inspecting the fidl
reference files needed by fidlcat on a per-package basis.

Testability: This is a refactor of a test with no new behaviors,
so having the existing tests pass is confirmation that the refactor
was successful.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381654
GitOrigin-RevId: a852b17e258bf60aaa90da9822b9dfddee262bf9
Change-Id: I0f7f844221a1c29f0da578b31621a9eb9cfd54f5