[roll] Roll fuchsia [gn sdk] Fixing fpave when device is in zedboot.
The ssh command to attempt to get the device in recovery mode
was failing if the device was already on the zedboot screen.

Since we tightened the error handling, this error caused the script
to exit.

Added a check for an error result on the ssh command so the script
continues.

Testability note: Cannot use multiply here since the test is run on the
sdk trybot, and via the test framework.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381906
GitOrigin-RevId: 6f15248e1d8f43f51344fd43897bcea380a2cbc7
Change-Id: I105dc2d6fb16a465167078a2c0ad1e893e3957c4