[roll] Roll experiences [settings] copies deprecated settings deps from topaz
This CL copies the deprecated setting mods' dependency (lib.settings) into the
experiences repo.

Deprecated settings code is being migrated to experiences so it can be removed
from the topaz repo (and eventually removed from experiences once new settings
feature parity is met).

Original-Bug: 45550

This is 2 of 4 CLs that need to be submitted in order:
1. [fxr/381099] add SDK visibility to experiences/settings
2. => [fxr/380992] copy lib.settings into experiences
3. [fxr/381113] delete lib.settings from topaz
4. [fxr/381106] remove SDK visibility from topaz

The set of CLs has been vaildated with:

        fx make-integration-patch

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/380992
GitOrigin-RevId: 2d88ab1870c25347a87c875392fe238abd55e1e6
Change-Id: Ia72c3a14b8c46de27ca6b9f429edcfca830c92ec