[roll] Roll fuchsia [mdns] Re-enable lsan for mdns_tests
I cannot repro the lsan failure in 1000 runs on a device nor
in the CQ. If the explanation at fxb/46754#c13 is correct, then
this might randomly come and go based on variations in the compiled
code. Alternately, this might have been fixed by some other change.
In any case I don't see any evidence to continue suppressing lsan
on this test.

Original-Fixed: 46754
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379078
GitOrigin-RevId: 0b7323a4f2a9f602872d99bad715d5f814229188
Change-Id: I44d2445b4b8c92984605ffa1aeda3aa95b28b86e