[roll] Roll fuchsia [audio_core] Re-enable lsan for audio_core_unittests
I cannot repro the lsan failure in 100 runs on a device (the test
was too slow to wait for 1000 runs) nor in the CQ. If the explanation
at fxb/46754#c13 is correct, then this might randomly come and go
based on variations in the compiled code. Alternately, this might
have been fixed by some other change. In any case I don't see any
evidence to continue suppressing lsan on this test.

Original-Fixed: 46533
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378855
GitOrigin-RevId: 15ed138174a50233919345c255bb89c08c9d1e51
Change-Id: I97a67b2fe763f8605f85f72070dde2151dc38c28