[roll] Roll fuchsia [bt][a2dp-source] Wait to connect after discovery.
When a peer is discovered, there is a possibility that it will become
the initiator and try to connect.  Wait two seconds before trying to
connect to newly discovered peers for this to happen.

Add LazyEntry functionality to DetachableMap to support lazy entry of
peer information and lazy retrieval of peers.

Original-Bug: 2783
Test: added wait_to_start_connection to bt-a2dp-source-tests
      added laxy_entry to bluetooth-create-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380148
GitOrigin-RevId: 55460ec42059d842c35a8c59d2345839c553c76a
Change-Id: Ib06546bc95a1035dac6802b00a89489664903e64