[roll] Roll fuchsia [workstation] Disable can_browse_web_test
The build rule for can_browse_web_test was accidentally broken in
December, and later the workstation product dropped support for
sessionctl add_mod, resulting in the test no longer working. The build
rule was recently fixed, causing test runs to fail as the test no longer
works. Therefore, we are disabling the test for the time being.

Original-Bug: 47690
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381696
GitOrigin-RevId: 78e35f22e73edd25f638ff56c4138e4b055ff34d
Change-Id: Id291e707265a2edc970c4629ccd7e838cb4860da