[roll] Roll fuchsia [zircon][scripts] Update gdb scripts to support python3.
GDB can be configured at build time to support Python2 or Python3,
with the latter more common since Python2 has been (long) deprecated.

Currently, the "zircon.elf-gdb.py" fails in Python3 due to its use of
"iteritems()" on a dict, which has been removed in Python3. Switch it
over to just "items()", which should work on both 2 and 3.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381899
GitOrigin-RevId: a32e4e6debd6f1eb0d4fdf2a22cf2c80395063b6
Change-Id: Ida49e8c3463e356aba5ea9d3a98309246b939c46