[roll] Roll fuchsia [sshutil] Use logger library for logging
The logger library associates a Context with a Logger object that
controls the logging level and sets a prefix for all logs printed by a
given tool. Botanist already uses it for other logs, so this will help
with consistency.

If there's no logger associated with the context, the library falls back
to using plain old stdlib log methods.

It's a bit weird for the keep-alive goroutines to take a context but not
check for ctx.Done(), but I wanted to keep this change small so I'll
address that in a follow-up.

Original-Bug: 48042
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382253
GitOrigin-RevId: b1cf19de34797bdb436aa014fa69d63834193cce
Change-Id: Ib3810743153630e3984417c87d7ede2953c63c16