[roll] Roll fuchsia [a11y] Adds an interface to a11y offer access to gestures.
This change adds interfaces that allow a11y to offer access to gestures. Because
a11y intercepts all gestures when it is on, this interfaces offer a way for a11y
to reserve some gestures, which can be listened by an UI (for example).

By listening to those gestures, an UI can perform special actions, and inform
a11y back of the result.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380303
GitOrigin-RevId: e27487a7b4aa60a20fc7000bce508e2e5a3f6232
Change-Id: I08a1bc925678bef0533b152e1272be6ee481b032