[roll] Roll fuchsia [media][tests] fix audio_device_tests race
In a similar fix to fxr/380981, this changes one of the
device gain test cases to register for an event BEFORE the
action that may trigger that event. This bug would not have
been causing flakes -- instead, if the race condition
occurred it would have hidden a valid bug.

MULTIPLY: `[
  {
    "name": "audio_device_tests",
    "total_runs": 20
  }
]`

Test: test-only code

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381168
GitOrigin-RevId: c1a320488b224b1d14e796d761030f76420c45dd
Change-Id: I14ad9d6872de174b7e8987723ce317c431524c52