[roll] Roll fuchsia [debugger] Move type finding to a separate file.
Previously resolving type forward definitions was in the EvalContext. It
didn't need to be there but it was a convenient place to put it. This
placement makes it difficult to use in FindName code because it creates
a circular dependency between FindName and EvalContext.

This patch moves out type forward-definition finding to a separate file,
and updates callers to use that. EvalContext can now provide a
FindNameContext for callers to conveniently use. This part of the change
is a simple code move.

Moving this out required updating some tests which relied on some
mocking capabilities of MockEvalContext that are now gone. In
particular, the VirtualBaseClassTestSetup class needs this. That test
harness now injects symbols into the index which makes things act more
like production code.

Mot using the MockEvalContext also made MockEvalContext's ability to set
the "should promote to derived" flag. Normally this is provided by the
"client" layer and is hooked up to the settings system. Since "expr"
library tests can't use this, a new TestEvalContextImpl is created for
"expr" library tests that need both a real EvalContextImpl and the
ability to control the "should promote to derived" setting.

This change will enable a future pass to implement resolving forward
type definitions in more places. But currently there should be no
behavior change.

Original-Bug: 50201

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381247
GitOrigin-RevId: 799fd75b93e1d2c6f8e6571b21e936507a226a1a
Change-Id: I125d2b38da71ea3534dfff47fdd3624384a25e1b