[roll] Roll fuchsia [debugger] Resolve forward declarations.
In some places around inheritance we were not handling type declarations,
instead expecting full definitions. These not call GetConcreteType as
needed.

GetConcreteType is moved out of EvalContext so it can be used by
FindName code. A helper function is left in EvalContext because this is
used in so many places and because calling it with the FindNameContext
from the EvalContext is annoying.

Make it clear that base classes can be non-concrete and audit all users
of InheritedFrom and VisitClassHierarchy to make them get concrete
versions of types when necessary.

Original-Bug: 50201

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381320
GitOrigin-RevId: d7e723ab0b73bf43907c4460840f7568f6e8271d
Change-Id: Ic67f024f71282c061d5216f36d09857f41515cb4