[roll] Roll fuchsia [lib/component] Specify dispatcher for app/context
...by calling context.BindStartupHandle once service registration is
done.

If a component or test didn't call fidl.Serve() before, it doesn't need
to call BindStartupHandle or create its own dispatcher now.

We enable service requests that spawn additional services to access
their dispatcher via fidl.Context, much like Marshaler and BindingKey.

This enables pipelined requests to work without the boilerplate of
manually plumbing through a dispatcher for the default behavior of "use
the same one this request was invoked on".

When updating other callers of fidl.Serve, take care not to introduce
additional dependencies (hence the varying strategies for handling
dispatcher creation errors).

Now that there are no callers of (*{{ .ServerName }}).AddWithCtx, remove
that method from the generated template code (and in some cases, remove
the "syscall/zx" import that it was the last consumer of).

Tested:
- fx build && fx qemu -kN 2>&1 > /dev/null & fx wait
- fx shell net stat show # exercises other out directories
- fx shell ls /hub/c/netstack.cmx/*/out
Original-Bug: 47468
Original-Bug: 48565

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/372874
GitOrigin-RevId: ba46a4ae73060a5b6e75886b31dd0b80ad3a2bf0
Change-Id: Iaa217bfbceba6aef6ccedf4179c3f77c327c3c7e