[roll] Roll fuchsia [fidlcat] implements ZX_INFO_HANDLE_TABLE
Add a flag to SyscallDecoderDispatcher to load info about the handles.

Optionally asks the debug agent to call zx_object_get_info with
ZX_INFO_HANDLE_TABLE for processes for which we need info about handles.

The goal here is to know the koid (immutable even if the handle moves
from one process to the other) for handles and the koid for the handle
peers.

This will be used later for graph generation: knowing the koid will
allow us to associate a handle we write on (in one process) to another
handle we read on (in another process).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382293
GitOrigin-RevId: 4a9d1cbcd327cf48bd3553e067998ada03d0f151
Change-Id: I42a92b70a0fc021af9b52b5bdf6acda34e355bac