[roll] Roll fuchsia [build] Restrict Zircon build API modules to default_deps
The consumers of the build API modules get confused if they describe
too many things that aren't actually being built.  So for now just
restrict the metadata collections that produce them to exactly the
graph of what's being requested by the legacy build integration.

Also prepare the Breakpad support for the toolchain runtimes to
continue to work right when the toolchain package delivers Breakpad
data directly in the future.

Original-Bug: 50337
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382393
GitOrigin-RevId: 6f707971fb3392dd25dca9885fa9f1f561463089
Change-Id: I910017c6e39e8750473d3b3652b7bb0cdfe7021b