[roll] Roll fuchsia [carnelian] Support user input in framebuffer mode
Added support for reading from
/dev/class/input_report for mouse, keyboard, and
touch input devices.

Since the scenic input structures are in flux, I
decided to create a set of input event structures
in Carnelian and translate both input report and
Scenic events to these new events.

Plumbing this new event structure through the
old, unused view modes made little sense, so this
change removes all view modes except Canvas and
Render. The notion of logical pixels was removed
along with the Scenic view mode. Carnelian might
someday re-introduce a density-independent
coordinate system it doesn't make much sense to
maintain what was there today.

Scenic reports all touch and mouse input as
pointer events, which does not strike me as a
well formed abstraction. Carnelian keeps these
two types of events separate while providing a
limited pointer facade for apps that want to
treat mouse events involving the primary button
as the same as touch events.

In documenting the new trait methods I gave setup
and update default implementations, so that no
one is required to implement them and thus
the documentation samples are easier to write
and maintain.

As part of porting terminal over to this even
model I added files neeeded to run it under fargo.

Testability: The handlers for Scenic and input
repot are designed to be tested in isolation and
a smattering of such tests are included in this
CL.

Testing: Passed all existing automated tests. Ran
all the Carnelian samples on a core and
workstation build. Used terminal on core and
workstation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375643
GitOrigin-RevId: f8a7265bfb234e31d56e4bf5010336b02f9f508a
Change-Id: I8dffff54106a82be08f06ceaff2f0c141d296278