[roll] Roll fuchsia [omaha-client] ComplexTime module
Time is complicated, especially when working with both monotonic and
wall time.  These attempt to make that easier.

ComplexTime has both a SystemTime and an Instant value.  These are not
necessarily the same moment.

PartialComplexTime can have either a SystemTime, an Instant, or both
and be a ComplexTime.

Both fully support all math operations that SystemTime and Instant do.

The Timer trait works with PartialComplexTime, to allow the flexible
setting of a wait based on either, or both, time sources.

The TimeSource trait is for producing ComplexTimes, and with the
MockTimeSource, allows for easy test mocking of time.

Original-Bug: 49956
Test: fx test omaha-client-lib-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381167
GitOrigin-RevId: 64fcb9f1fa2d22b228f1ff4ec3be2c8da93167a7
Change-Id: Id6cca16cd41e40e6e226ca1cd1fe7780555222b2