[roll] Roll fuchsia [FDF] Add a comment about deprecated api
fuchsia.device.manager.administrator is deprecated and moved to fuchsia.hardware.power.statecontrol.admin.fidl.
fuchsia.device.manager has many fidl apis that are internal to device manager and are not appropriate to be exposed to SDK.
For this reason a new api fuchsia.hardware.power.statecontrol.admin.fidl
is created with the same functionality and is exposed to SDK.

The new api in future will be implemented by system power manager because the system state transitions do not belong in
device manager.Once all the clients are moved to the new api, device.manager.administrator.fidl will be removed.

Original-Bug:42257

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382700
GitOrigin-RevId: ce3799e0e26b6486b698dd2c2d39f6c441f3a126
Change-Id: I93beab796bff1083f17e659538f0e5ea0767d3d0