[roll] Roll fuchsia [audio] Add option for extending audio_core driver interface
Extend the AddDeviceByChannel FIDL API and various internal audio_core
interfaces with overloads to differentiate instances of the
driver interface supporting the legacy manual serializations versus
the FIDL based interface. Subsequent CLs add the actual support for
the FIDL interface and all unit tests are kept with the legacy interface.

We also move SelectBestFormat as a driver method, this avoids users
spefifying format_ranges() that returns a type specify to the legacy
implementation.

In order to overload AddDeviceByChannel FIDL definition we need to
import the audio driver FIDL definitions, due to this we need to flag
the audio driver interface as sdk_category = partner same as
audio_core's FIDL.

Test: fx run-test audio_core_unittests
Test: fx run-test audio_device_service_tests
Test: E2E audio input and output on Astro
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376855
GitOrigin-RevId: 125a526811fa750d2abd58a2ec335d85fc95df10
Change-Id: Ie56001fb4200962ec7ac70d3a635793611e9efac