[roll] Roll fuchsia [audio_core] Separate readable from writable streams
Currently we are not consistently pairing each LockBuffer with an
UnlockBuffer. For example, TapStage never unlocks tap_ and MixStage
never unlocks output_stream_. These bugs don't cause actual problems at
the moment because UnlockBuffer is a nop for all streams except for
PacketQueue. Nonetheless we should fix these bugs.

In looking into a fix, it occured to me that Stream is used in two
different ways: as an input (a reader) and as an output (a writer). The
UnlockBuffer(release_buffer) API doesn't make sense for writers because
writers don't have anything to "release" -- UnlockBuffer() is
sufficient. With that in mind, this CL splits Stream into three classes:

BaseStream
  +-- Stream (ReadLock, ReadUnlock)
  +-- WritableStream (WriteLock, WriteUnlock)

Stream describes read-only streams while WritableStream describes
write-only streams. I decided not to rename Stream to ReadableStream
because that would result in a lot more code churn. I also split
RingBuffer into three classes:

BaseRingBuffer
  +-- ReadableRingBuffer (inherits from Stream)
  +-- WritableRingBuffer (inherits from WritableStream)

This is (hopefully!) a simplification of RingBuffer -- it's now obvious
from the type whether the buffer is used as an input or output, as
opposed to the prior implementation which required a dynamic check of
the Endpoint() method.

There are just two implementations of WritableStream: IntermediateBuffer
and WritableRingBuffer.

This is a pure refactoring. There are no behavior changes. A future CL
will add the missing unlocks.

Original-Bug: 50442
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382078
GitOrigin-RevId: 954dde95fa90899821945bdcfecda6c5755389ae
Change-Id: Ia8b0e0030e897c6743f56fc5e16b16ebf4eac210