[roll] Roll fuchsia [swd-e2e-tests] Rewrite run-emu to sync with `fx qemu`
The e2e tests `run-emu` scripts were originally were a prototype to get
an update-able instance of QEMU. Since written, it didn't keep up with
`fx qemu`, and so it never implemented the rest of the features
supported by `fx qemu`, nor some of the bug fixes it has received.
Specifically, the networking configuration doesn't work correctly for
some workstatsions.

In order to install the SSH keys onto the QEMU target we bootstrap the
virtual device by paving it. The script constructs an image that will
boot to zedboot and runs the paver.

Original-Fixed: 49125
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381274
GitOrigin-RevId: 9e00fe2ae62a5637740a1451446b4f582ca1d14c
Change-Id: Ie6a111722f40c36d29e02cf6fb3bbfa746309181