[roll] Roll fuchsia [dhcp] Send response through the server's IP
Without this change, the server sends responses from the socket bound
to the ANY address. There is no way for the netstack to determine which
NIC these responses should go out of, so the default route is always
used. This does not work when multiple NICs are present and are
connected to different networks. In this case, responses that are
supposed to go out of non-default NICs are always routed through the
default one, as if they are dropped by the netstack.

Test:
fx test -od netstack_fidl_integration_test -- acquire_dhcp
fx test -od dhcp_tests

Original-Bug: 48027

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380533
GitOrigin-RevId: 6aab6bff6f225573a1619c572424d50817524e71
Change-Id: I25477d91d72b810c0da575b87145bf2d9a22690a