[roll] Roll fuchsia [rfc] Platform versioning
This document introduces the notion of an *API level* and an *ABI
revision* to the Fuchsia platform. End-developers build against a *target
API level*, which determines which declarations are visible to the
application. The *target API level* also becomes embedded in the
compiled application as a *target ABI revision*, which indicates the
semantics the application expects from the platform. A given release of
the Fuchsia platform typically supports multiple ABI revisions, which
lets the platform run older applications while still providing a path
for evolving the platform.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376759
GitOrigin-RevId: 30c3af21d996abe989563359fef83bc8c9d76664
Change-Id: I9d7f1825d15e4329e0a27269a2950b25678515f7