[roll] Roll fuchsia [audio] Add audio_core drivers FIDL interface
Adds to audio_core support for the FIDL audio driver API using
the HLCPP FIDL bindings. The manual serialization support is kept,
and it is chosen if defvs lists /dev/class/audio-input instead of
/dev/class/audio-input-2 (similarly for output).
Manual serialization support is renamed to AudioDriverV1 and we
add a AudioDriverV2 class supporting FIDL. Note that we'll remove
AudioDriverV1 once the transition is completed.
Tests are kept for the V1 interface and new ones added for V2
(also a FakeAudioDriverV2 for V2).

Test: fx run-test audio_core_unittests
Test: fx run-test audio_device_service_tests
Test: fx run-test audio_fidl_tests
Test: fx run-test audio_pipeline_tests
Test: fx run-test audio_device_tests on Astro with and without
transition of SimpleAudio to V2
Test: fx run-test audio_core_hardware_tests on Astro with and without
transition of SimpleAudio to V2
Test: E2E audio input and output on Astro with and without transition of
SimpleAudio to V2

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376854
GitOrigin-RevId: f5bad2c41e70d1227c0c6d23f50f577a59cab9b9
Change-Id: I1c4dffadd3d787d1921a9aa73499744d644aca30