[roll] Roll fuchsia [audio_core] Fix device startup race
We need to ensure we don't form any links to an AudioOutput until the
device has been activated. To do so, we move the ActivateSelf call later
in the device startup process, and also fix an issue where a plug
detect message could cause a device to be linked before it's been
activated.

Fixes: 47251
Test: fx test audio_core_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382690
GitOrigin-RevId: 1cc137139bc6b2a4c362086de5e3c83bd0192a35
Change-Id: Ibe456d88ccbdb9f05d6a684c02cf99bf18731a60