[roll] Roll fuchsia [astro][audio] Fix high clockrates
Recent evaluation of drivers to support 96kHz frame rate revealed
that the audio clock tree has a maximum input frequency of 1GHz.
In order to get the output of HIFI_PLL under this limit while still
being able to hit available division ratios, the tdm framing to the
codec had to be changed to 64 sclk per frame (was 128).  This
cascaded into an undocumented issue with the tas2770 where it does
not operated properly with (L+R)/2 mixing and 64sclks per frame.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381884
GitOrigin-RevId: 46969c810eba5bbf1a4aa1f9b9b4f5ed02cf67d4
Change-Id: I7f8bef5810ab9817a193b57bb2743cf744ed64ee