[roll] Roll fuchsia [feedback] Simplify AnnotationsProviders
After revisiting the annotation collection process I realized that it's
overly complex for what it needs to achieve. We were doing a ton of
unncessary and redundant filtering that made the process hard to
understand and reason about.

Instead of filtering the allowlist prior to instantiating each provider,
we just pass the allowlist to a provider it gets its annotations and
rely on the provider to do the filtering.

Additionally, we always instantiate all annotation providers instead of
instantiating the providers that have a non-empty intersection with the
allowlist. This reduces the complexity of annotation collection at the
cost of putting the burden of not collectiong annotations when the
intersection is empty on the providers

TESTED=`fx test feedback-data-tests`

Original-Bug: 49922

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382692
GitOrigin-RevId: 27f029a124df0f33adc22c29335063c3a97ff6f1
Change-Id: I0e84507bc5a3084cd14a0c0600b9e468e2039c37