[roll] Roll fuchsia [trace] support triggers (alert:action mapping)
This CL modifies the 'trace' util to implement triggers. A new
command line option is added to specify triggers, and the json 'spec'
allows triggers to specified.

TEST: fx test trace_tests
      Added unit tests for spec changes.
      Added a new integration test for triggers.

The new test provider runs for an hour, so that the alert is needed
to end the test. It also sets the session duration to an hour. A
trigger is specified to stop the session. Because triggers work, the
test terminates promptly. If triggers didn't work, the test would run
for one hour, which would cause the test runner to time out and
declare the test failed.

Original-Bug: 49629

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382766
GitOrigin-RevId: 0337c742c307d9a634e8653a3bc397d0ae0ef1a0
Change-Id: I3b743192181ca1e88a7c45463bab73d082038666