[roll] Roll fuchsia [cobalt] Changes variable name to |validated_clock|.
This came up while attempting to understand some code.
Within cobalt_app.cc there is an instance of FuchsiaSystemClock
which implements the interface ValidatedClock. A pointer to this
is passed to CreateCobaltConfig where it is used as the value
of the |validated_clock| field. There are also instances of
std::chrono::system_clock and util::SystemClock used in the
same class. There was (prior to my change) a variable called
system_clock of type util::SystemClock even though there was
also a parameter called system_clock of type FuchsiaSystemClock.

I found all of this difficult to follow. I think it is a little
easier to read if the instances of FuchsiaSystemClock are called
|validated_clock|.

I also changed a log statement in the test_app.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383335
GitOrigin-RevId: 05319641230b9c96229da8b79f631ee96e7f7b9a
Change-Id: I6cdd9c7e3b1c2bab5775311364f6d96b775005f9