[roll] Roll fuchsia [camera] Add stream orphaning test.
This change adds a new test that verifies that Stream clients can remain
connected even after the connection to its parent Device is closed. This
can be used to cede exclusive access to a device, opportunistically
remaining connected, but permitting a new client to take ownership
without coordination.

This change also fixes an issue in the test suite Sync method that
prevented calling it on a StreamPtr more than once in the object's
lifetime.

MULTIPLY: `[
  {
    "name": "device_test",
    "total_runs": 50
  }
]`

Test: added new test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383150
GitOrigin-RevId: b763ceb7c446e76883f6645f75333b513f19fc5d
Change-Id: I8cab6665cb1cee9e1f5b6a1375c87f3453797132