[roll] Roll fuchsia [storage] Clang-format blobfs/include
No code change.

The main change is in format.h which wass still 4-space indented for
most of it due to clang-format being disabled. Keeps clang-format=off
around some constants at the top that seemed a better fit with the
tabular formatting. Reformats the rest.

Re-wrap some longer comments to 100 cols that used the older standard of
80.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382918
GitOrigin-RevId: fd39552c6b1f35afd930b9d4664c6bc6eebbcbaf
Change-Id: I2514a6efa60513344a47d12d4577ee7e92f78be7