[roll] Roll fuchsia [runtests] Combine all profile sinks in memory before writing them
Components that spawn large number of subprocesses could produce a lot
of data sink dumps for the same module. These would all be merged to a
single file, but constantly reading, merging and writing the same file
is expensive. So instead, we merge all profiles for the same module in
memory before writing it out to disk.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379824
GitOrigin-RevId: 7410f82d4f48922ab76a904c0ddba55dd5a918fe
Change-Id: I25dbd585bf90872be1f3fa31749b16357b5385e1