[roll] Roll fuchsia [identity][modular] Unhook Fuchsia Identity from modular, step 7.
Step 7 removes account ID as a parameter across the user provider
surface, since session user provider is now deducing account ID from the
existing directories. This is implemented by adding a new method to the
UserProvider FIDL, marked transitional to allow a soft transition since
this interface is in the SDK.

While in the API I updated the documentation for all the non functional
methods and marked them transitional so we can begin the graceful
removal.

Testing: Tested the setup flow manually on a test device.

Original-Bug: 45946,48782
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382895
GitOrigin-RevId: e36ceb32ad59df652164ecbd1672af7e12f58072
Change-Id: I3924a3e6dd7c7fb26784cdb95619796c9c6dc054