[roll] Roll fuchsia [audio_core][tests] Mixer-creation profiling
New test cases measure the elapsed time required to create audio
mixer objects, with various parameters. This demonstrates that the
significant cost of creating SincSamplers is in the filter creation,
which scales linearly with the width of the filter. This width is
directly linked to the down-sample ratio, with a lower bound of the
width of the filter at unity rate-conversion. Also demonstrated is
the reality that the Point and Linear samplers have no such issue
(as they use no such filter).

"run_test_component audio_fidelity_tests --profile" shows all details

Test: test-only change; no production impact

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382938
GitOrigin-RevId: 51f61da821bfc770801dd1747e5050982e0cd8b1
Change-Id: I43df7c6c0678aa1626b695df8ae113115c6da8cb